titanic/requirements.txt
konan-sdk>=1.3.0
# Add your python requirements here
pyYaml
pandas
sklearn
